% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{clear_individual_redcap_instruments}
\alias{clear_individual_redcap_instruments}
\title{Clears individual REDCap instrument data frames, if they are present.}
\usage{
clear_individual_redcap_instruments()
}
\description{
Clears individual REDCap instrument data frames, if they are present.
}
