% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{hash_arc_number}
\alias{hash_arc_number}
\title{Creates a version of the data in which the specified \code{id_field} is replaced by a hashed version.}
\usage{
hash_arc_number(data, id_field)
}
\description{
Creates a version of the data in which the specified \code{id_field} is replaced by a hashed version.
}
