% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{load_and_prep_cdw_data}
\alias{load_and_prep_cdw_data}
\title{Read CDW csv files and apply standard preparations.
Probably not needed by users.}
\usage{
load_and_prep_cdw_data(
  filename,
  base_filename,
  protocol_has_wildcard = F,
  hide_arc_number = T
)
}
\arguments{
\item{filename}{The name of the file to read.}

\item{base_filename}{The file location.}

\item{hide_arc_number}{If \code{TRUE}, hashes the ARC #. Default: \code{TRUE}.}
}
\description{
Read CDW csv files and apply standard preparations.
Probably not needed by users.
}
