% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{load_cdw_data}
\alias{load_cdw_data}
\title{Load all CDW data from nightly dump files and combine into a single data frame.}
\usage{
load_cdw_data(
  protocol_names,
  path,
  keep_individual_instruments = T,
  hide_arc_number = T
)
}
\arguments{
\item{protocol_names}{The names of the protocol to load. This is the first portion of the file name. You can specify multiple names in a vector to load and combine multiple files.}

\item{path}{The file path where the nightly dump files are stored.}

\item{keep_individual_instruments}{If \code{TRUE}, will also load each individual CDW file into its own data frame in your R session. Default: \code{TRUE}.}

\item{hide_arc_number}{If \code{TRUE}, hashes the ARC #. Default: \code{TRUE}.}
}
\value{
The combined CDW data in a long-format data frame.
}
\description{
Load all CDW data from nightly dump files and combine into a single data frame.
}
