% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{load_individual_redcap_instruments}
\alias{load_individual_redcap_instruments}
\title{Load individiual REDCap instruments into the R environment}
\usage{
load_individual_redcap_instruments(
  API_key,
  arc_number_field_name,
  complete_redcap_data = NULL,
  redcap_server = c("redcapsurvey", "redcap"),
  envir = NULL
)
}
\arguments{
\item{API_key}{The REDCap API token.}

\item{arc_number_field_name}{The name of the ARC # field in your REDCap project.}

\item{complete_redcap_data}{If you have already loaded the complete REDCap data using \code{load_redcap_instruments()}, you can make this function faster by providing that data frame here. This is not required.}

\item{redcap_server}{Either \code{'redcapsurvey'} or \code{'redcap'}. Default: \code{'redcapsurvey'}.}
}
\description{
This may load a large number of data frames into your R session. To clear these individual data frames, call \code{clear_individual_redcap_instruments()}.
}
