% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{load_redcap_instruments}
\alias{load_redcap_instruments}
\title{Load all REDCap instruments into a single large data frame}
\usage{
load_redcap_instruments(
  API_key,
  arc_number_field_name,
  redcap_server = c("redcapsurvey", "redcap"),
  raw_or_label = c("raw", "label"),
  keep_individual_instruments = F,
  hide_arc_number = T
)
}
\arguments{
\item{API_key}{The REDCap API token.}

\item{arc_number_field_name}{The name of the ARC # field in your REDCap project.}

\item{redcap_server}{Either \code{'redcapsurvey'} or \code{'redcap'}. Default: \code{'redcapsurvey'}.}

\item{raw_or_label}{Either \code{'raw'}, to keep the raw response values in multiple choice questions, or \code{'label'} to keep the corresponding labels. Default: \code{'raw'}.}

\item{keep_individual_instruments}{If \code{TRUE}, will also load each individual REDCap instrument into its own data frame in your R session. Default: \code{FALSE}.}
}
\value{
A wide-format data frame containing data from all REDCap instruments.
}
\description{
Load all REDCap instruments into a single large data frame
}
