% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{redcap_cdw_id_mapping}
\alias{redcap_cdw_id_mapping}
\title{Reveal the mapping between REDCap ID, CDW ARC #, and hashed (hidden) ARC #.}
\usage{
redcap_cdw_id_mapping()
}
\value{
The ID mapping
}
\description{
Note that the mapping shown will be based on the most recent REDCap data loaded with \code{load_redcap_instruments()} or \code{load_individual_redcap_instruments()}.
}
