% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{combine_redcap_and_cdw_data}
\alias{combine_redcap_and_cdw_data}
\title{Combine the REDCap and CDW data into a single long format data frame.}
\usage{
combine_redcap_and_cdw_data(redcap_data, cdw_data)
}
\arguments{
\item{instrument_env}{The environment containing individual REDCap instruments. Default: .InstrumentEnv}
}
\value{
The fully merged REDCap and CDW data in long format.
}
\description{
Combine the REDCap and CDW data into a single long format data frame.
}
